removes projector logic
