--- src/main/java/sc/fiji/bdvpg/bdv/projector/AccumulateMixedProjectorARGBFactory.java
@@ -1,63 +1,0 @@
-/*-
- * #%L
- * BigDataViewer-Playground
- * %%
- * Copyright (C) 2019 - 2021 Nicolas Chiaruttini, EPFL - Robert Haase, MPI CBG - Christian Tischer, EMBL
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-package sc.fiji.bdvpg.bdv.projector;
-
-import bdv.viewer.SourceAndConverter;
-import bdv.viewer.render.AccumulateProjectorFactory;
-import bdv.viewer.render.VolatileProjector;
-import net.imglib2.RandomAccessible;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.type.numeric.ARGBType;
-
-import java.util.List;
-import java.util.concurrent.ExecutorService;
-
-/**
- * For information about this projector, see {@link AccumulateMixedProjectorARGB}
- */
-
-public class AccumulateMixedProjectorARGBFactory implements AccumulateProjectorFactory< ARGBType >
-{
-	public VolatileProjector createProjector(
-			List< VolatileProjector > sourceProjectors,
-			List<SourceAndConverter< ? >> sources,
-			List< ? extends RandomAccessible< ? extends ARGBType > > sourceScreenImages,
-			RandomAccessibleInterval< ARGBType > targetScreenImage,
-			int numThreads,
-			ExecutorService executorService )
-	{
-		return new AccumulateMixedProjectorARGB(
-						sourceProjectors,
-						sources,
-						sourceScreenImages,
-						targetScreenImage,
-						numThreads,
-						executorService );
-	}
-}
